split processor.h to generator.h and filter.h
