+++ beeps/src/oscillator.cpp
@@ -1,4 +1,4 @@
-#include "processor.h"
+#include "beeps/generator.h"
 
 
 #include "SineWave.h"
