+++ beeps/src/time_stretch.cpp
@@ -1,4 +1,4 @@
-#include "processor.h"
+#include "beeps/filter.h"
 
 
 #include "signalsmith-stretch.h"
